% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_index_from_FASTA.R
\name{FM_index_from_FASTA}
\alias{FM_index_from_FASTA}
\title{FM_index_from_FASTA}
\usage{
FM_index_from_FASTA(input, output, save = TRUE)
}
\arguments{
\item{input}{path to the FASTA file}

\item{output}{path to the folder in which to save the data structures}

\item{save}{default = TRUE, can be modified to avoid saving data in the output path as separate files}
}
\value{
Object of type FM_index (list-like object)
}
\description{
Generates FM_index structure starting from an input FASTA file
}
\details{
FM index is composed of 4 different elements:
\itemize{
\item Suffix Array --> stored as a Dataframe, contains all suffixes and their indexes of the original string
\item Burrows-Wheeler Transform --> a transformed string
\item Occurrency matrix --> Dataframe storing the cumulative count table of characters in the BWT
\item Count array --> Dataframe containing, for each character, the count of characters that are lexicographically smaller in the sorted BWT
}

The FASTA file is parsed using a Biostrings function (readDNAStringSet) in order to ensure a robust and reliable parsing method
The FM Index returned by the function contains also the header of the FASTA file in order to keep informations related to the original sequence
}
\examples{
# example creation of an FM index from a FASTA file
FM_index <- FM_index_from_FASTA(system.file("extdata", "NM_001185098.2", package = "FMIndexR"))
}
