% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_TUB.R
\docType{data}
\name{FM_TUB}
\alias{FM_TUB}
\title{FM_TUB}
\format{
An object of class FM_index, containing:
\describe{
\item{Suffix Array}{Description of the Suffix Array}
\item{BWT}{Description of the Burrows-Wheeler Transform (BWT)}
\item{Occ matrix}{Description of the Occurrence matrix}
\item{Count array}{Description of the Count array}
\item{FASTA header}{Description of the FASTA header}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/nuccore/NM_177972.3}
}
\usage{
data(FM_TUB)
}
\description{
An example FM index created from a FASTA file.
}
\details{
The FASTA file is stored in inst/extdata/TUB_FASTA.txt
}
\examples{
data(FM_TUB)
}
\keyword{datasets}
