% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BackwardSearch.R
\name{BackwardSearch}
\alias{BackwardSearch}
\title{BackwardSearch}
\usage{
BackwardSearch(FM_index, pattern, store_elems = FALSE)
}
\arguments{
\item{FM_index}{object of class FM_index obtained using FM_index_from_FASTA}

\item{pattern}{string containing the pattern to look for}

\item{store_elems}{default = FALSE, can be modified to save the original sequence, the pattern indexes and the pattern in case a manual check wants to be made}
}
\value{
Prints a report of the number of patterns and their location, if store_elems = TRUE a report for manual check can be returned
}
\description{
Looks for patterns inside the original sequence, using its FM_index and performing a backward search procedure for finding patterns
}
\examples{
# example creation of pattern search using BackwardSearch and an FM Index
FM_index <- FM_index_from_FASTA(system.file("extdata", "NM_001185098.2.txt", package = "FMIndexR"))
BackwardSearch(FM_index,"ATC")
}
