% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FM_TUB.R
\docType{data}
\name{FM_TUB}
\alias{FM_TUB}
\title{FM_TUB}
\format{
An object of class FM_index, containing:
\describe{
\item{Suffix Array}{The Suffix Array of the sequence}
\item{BWT}{The Burrows-Wheeler Transform (BWT) of the sequence}
\item{Occ matrix}{The Occurrencies matrix of the BWT}
\item{Count array}{The Count array of the sequence}
\item{FASTA header}{The FASTA header of the sequence}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/nuccore/NM_177972.3}
}
\usage{
data(FM_TUB)
}
\description{
An example FM index created from a FASTA file.
}
\details{
The FASTA file is stored in inst/extdata/TUB_FASTA.txt
}
\examples{
data(FM_TUB)
}
\keyword{datasets}
